refactor: extract shared segment validation, add v2 debug logging
- Extract validate_and_reorder_segments() to eliminate v1/v2 duplication
- Widen SegmentGap error fields from u16 to u32 for v1+v2 compatibility
- Replace magic numbers with named constants (MAX_SECTION_DATA_SIZE, DEFAULT_V2_CHUNK_SIZE)
- Add log::debug! to v2 reader path (device_info, sector_table, md5_hash)
- Make ewf2 module pub(crate) — internal implementation detail

+++ ewf/src/error.rs
@@ -23,7 +23,7 @@ pub enum EwfError {
     Decompression(String),
 
     #[error("segment gap: expected segment {expected}, got {got}")]
-    SegmentGap { expected: u16, got: u16 },
+    SegmentGap { expected: u32, got: u32 },
 
     #[error("no segment files found matching: {0}")]
     NoSegments(String),
